# docs/api/openapi-v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raychi content and site API",
    "version": "0.2.0",
    "description": "v0.2 extends v0.1 without removing legacy article endpoints."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/public/articles": {
      "get": {
        "operationId": "listPublicArticles",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "已发布文章摘要",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticlePage"
                }
              }
            }
          }
        }
      }
    },
    "/public/articles/{slug}": {
      "get": {
        "operationId": "getPublicArticle",
        "parameters": [
          {
            "$ref": "#/components/parameters/Slug"
          }
        ],
        "responses": {
          "200": {
            "description": "最新发布快照",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticle"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/articles": {
      "get": {
        "operationId": "listAdminArticles",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "工作稿列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminArticlePage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createArticle",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新建工作稿",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminArticle"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/articles/{id}": {
      "get": {
        "operationId": "getAdminArticle",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleId"
          }
        ],
        "responses": {
          "200": {
            "description": "工作稿详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminArticle"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "saveArticle",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新工作稿，公开快照不变",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminArticle"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/articles/{id}/publish": {
      "post": {
        "operationId": "publishArticle",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "将工作稿复制为公开快照",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminArticle"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/articles/{id}/unpublish": {
      "post": {
        "operationId": "unpublishArticle",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "撤回文章及当前公开图片",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminArticle"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/public/contents": {
      "get": {
        "operationId": "listPublicContents",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ARTICLE",
                "POST",
                "THOUGHT"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published snapshots only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContentPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/public/contents/{type}/{slug}": {
      "get": {
        "operationId": "getPublicContent",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ARTICLE",
                "POST",
                "THOUGHT"
              ]
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContent"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/contents": {
      "get": {
        "operationId": "listAdminContents",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ARTICLE",
                "POST",
                "THOUGHT"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Working copies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createContent",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ARTICLE",
                "POST",
                "THOUGHT"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/contents/{id}": {
      "get": {
        "operationId": "getAdminContent",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Working copy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "saveContent",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved working copy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/contents/{id}/publish": {
      "post": {
        "operationId": "publishContent",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Publish content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/contents/{id}/unpublish": {
      "post": {
        "operationId": "unpublishContent",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unpublish content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "RAYCHI_SESSION"
      }
    },
    "parameters": {
      "ArticleId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "Slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50,
          "default": 10
        }
      },
      "CsrfHeader": {
        "name": "X-CSRF-TOKEN",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Error": {
        "description": "统一错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ArticleInput": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          },
          "title": {
            "type": "string",
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "maxLength": 600
          },
          "bodyMarkdown": {
            "type": "string",
            "maxLength": 1000000
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "string",
              "maxLength": 40
            }
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1000
          }
        },
        "description": "创建时各字段可省略；保存时 version 必填，未给字段将按服务端规则清空或保留。"
      },
      "VersionInput": {
        "type": "object",
        "required": [
          "expectedVersion"
        ],
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AdminArticle": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "status",
          "version",
          "title",
          "summary",
          "bodyMarkdown",
          "tags",
          "hasUnpublishedChanges",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "bodyMarkdown": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasUnpublishedChanges": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "publicUpdatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "PublicArticle": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "tags",
          "publishedAt",
          "publicUpdatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "bodyMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "description": "列表为 null，详情为 Markdown 正文"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publicUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminArticlePage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminArticle"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PublicArticlePage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicArticle"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ContentInput": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          },
          "title": {
            "type": "string",
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "maxLength": 600
          },
          "bodyMarkdown": {
            "type": "string",
            "maxLength": 1000000
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "string",
              "maxLength": 40
            }
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1000
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "description": "创建时各字段可省略；保存时 version 必填，未给字段将按服务端规则清空或保留。"
      },
      "AdminContent": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "status",
          "version",
          "title",
          "summary",
          "bodyMarkdown",
          "tags",
          "hasUnpublishedChanges",
          "createdAt",
          "updatedAt",
          "type",
          "category"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "bodyMarkdown": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasUnpublishedChanges": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "publicUpdatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": [
              "ARTICLE",
              "POST",
              "THOUGHT"
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "PublicContent": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "tags",
          "publishedAt",
          "publicUpdatedAt",
          "type",
          "category"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "bodyMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "description": "列表为 null，详情为 Markdown 正文"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publicUpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": [
              "ARTICLE",
              "POST",
              "THOUGHT"
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "AdminContentPage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminContent"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PublicContentPage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicContent"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}